+++ Taller/NutMatrix.h
@@ -1,6 +1,6 @@
 #ifndef NUTMATRIX_H_INCLUDED
 #define NUTMATRIX_H_INCLUDED
 
-void NutMatrix(double Mjd_TT, double matrizNutacion[3][3]);
+void NutMatrix(long double Mjd_TT, long double matrizNutacion[3][3]);
 
 #endif // NUTMATRIX_H_INCLUDED
