+++ Taller/R_z.c
@@ -2,10 +2,10 @@
 #include "R_z.h"
 #include "MatlabUtilities.h"
 
-void R_z(double angle, double matrizRotada[3][3])
+void R_z(long double angle, long double matrizRotada[3][3])
 {
-    double c = cos(angle);
-    double s = sin(angle);
+    long double c = cos(angle);
+    long double s = sin(angle);
 
     zeros(3,3, matrizRotada);
