+++ Taller/lambert_gooding.c
@@ -2,27 +2,27 @@
 #include "lambert_gooding.h"
 #include <math.h>
 
-void lambert_gooding(double r1,double r2, double tof, double mu, bool long_way, int multi_revs, double v1[3], double v2[3])
+void lambert_gooding(long double r1,long double r2, long double tof, long double mu, bool long_way, int multi_revs, long double v1[3], long double v2[3])
 {
 
 }
 
-void vlamb(double gm,double r1[3], double r2[3] , double th, double tdelt,  double salida_n_vri_vti_vrf_vtf[5])
+void vlamb(long double gm,long double r1[3], long double r2[3] , long double th, long double tdelt,  long double salida_n_vri_vti_vrf_vtf[5])
 {
 
 }
 
-void tlamb(double m,double q,double qsqfm1, double x, double n, double salida_t_dt_d2t_d3t[4])
+void tlamb(long double m,long double q,long double qsqfm1, long double x, long double n, long double salida_t_dt_d2t_d3t[4])
 {
 
 }
 
-double d8rt(double x)
+long double d8rt(long double x)
 {
     return sqrt(sqrt(sqrt(x)));
 }
 
-void xlamb (double m, double q, double qsqfm1, double tin, double salida_n_x_xpl[3])
+void xlamb (long double m, long double q, long double qsqfm1, long double tin, long double salida_n_x_xpl[3])
 {
 
 }
